+++ server/python/app/storage/cosmos_db_conversation_store.py
@@ -4,34 +4,16 @@ from azure.cosmos import PartitionKey
 from azure.cosmos.aio import CosmosClient
 from azure.cosmos.exceptions import CosmosResourceNotFoundError
 
-from .identity import get_azure_credential_async
-from .models import Conversation
+from ..models import Conversation
+from ..utils.identity import get_azure_credential_async
+from .conversation_store import ConversationStore
 
 
-class ConversationsStore:
+class CosmosDBConversationStore(ConversationStore):
     """
-    Interface for storing and retrieving conversations.
-    This is an abstract class that defines the methods for
-    getting, setting, deleting, and listing conversations.
+    CosmosDB implementation of the ConversationStore interface.
     """
 
-    async def get(self, conversation_id: str) -> Conversation | None:
-        raise NotImplementedError
-
-    async def set(self, conversation: Conversation):
-        raise NotImplementedError
-
-    async def delete(self, conversation_id: str):
-        raise NotImplementedError
-
-    async def list(self) -> list[Conversation]:
-        raise NotImplementedError
-
-    async def get_by_session_id(self, session_id: str) -> Conversation | None:
-        raise NotImplementedError
-
-
-class CosmosDBConversationsStore(ConversationsStore):
     def __init__(self):
         if endpoint := os.getenv("AZURE_COSMOSDB_ENDPOINT"):
             self.client = CosmosClient(
@@ -85,9 +67,6 @@ class CosmosDBConversationsStore(ConversationsStore):
     async def set(self, conversation: Conversation):
         container = await self._get_container()
         data = conversation.model_dump()
-
-        # TODO use patch instead of upsert for certain operations (e.g. transcript/rtt)
-        # to avoid overwriting the entire document
         await container.upsert_item(data)
 
     async def delete(self, conversation_id: str) -> None:
@@ -98,7 +77,6 @@ class CosmosDBConversationsStore(ConversationsStore):
         container = await self._get_container()
         query = "SELECT * FROM c"
         items = container.query_items(query)
-
         return [Conversation(**item) async for item in items]
 
     async def get_by_session_id(self, session_id: str) -> Conversation | None:
@@ -106,46 +84,6 @@ class CosmosDBConversationsStore(ConversationsStore):
         query = "SELECT * FROM c WHERE c.session_id = @session_id"
         params = [{"name": "@session_id", "value": session_id}]
         items = container.query_items(query, parameters=params)
-
         async for item in items:
             return Conversation(**item)
-
         return None
-
-
-class InMemoryConversationsStore(ConversationsStore):
-    def __init__(self):
-        self._store = {}
-
-    async def get(self, conversation_id: str) -> Conversation | None:
-        return self._store.get(conversation_id)
-
-    async def set(self, conversation: Conversation):
-        self._store[conversation.id] = conversation
-
-    async def delete(self, conversation_id: str):
-        if conversation_id in self._store:
-            del self._store[conversation_id]
-
-    async def list(self) -> list[Conversation]:
-        return list(self._store.values())
-
-    async def get_by_session_id(self, session_id: str) -> Conversation | None:
-        for conversation in self._store.values():
-            if conversation.session_id == session_id:
-                return conversation
-
-        return None
-
-
-def get_conversations_store():
-    """
-    Factory to select CosmosDB or in-memory store based on environment variables.
-    """
-    if os.environ.get("AZURE_COSMOSDB_ENDPOINT") or os.environ.get(
-        "AZURE_COSMOSDB_CONNECTION_STRING"
-    ):
-        return CosmosDBConversationsStore()
-
-    # Fallback to in-memory store if no CosmosDB credentials are provided
-    return InMemoryConversationsStore()
